refactor github actions into object
