+++ lib/event/create-update-event.js
@@ -1,8 +1,6 @@
 var config = require('../../common/config');
-var createEventFile = require('../github/create-event-file');
 var debug = require('debug')('gitevents-events');
-var getEventFile = require('../github/get-event-file');
-var updateEventFile = require('../github/update-event-file');
+var githubHelper = require('../github-helper');
 
 module.exports = function createUpdateEvent(event, github) {
   return new Promise(function(resolve, reject) {
@@ -13,7 +11,7 @@ module.exports = function createUpdateEvent(event, github) {
       name: event.name
     };
 
-    getEventFile(github, config, commit.filename)
+    githubHelper(github).getEventFile(commit.filename)
       .then(function(eventsFile) {
         var contents = new Buffer(eventsFile.content, eventsFile.encoding).toString('utf8');
         var previousEvent = JSON.parse(contents);
@@ -23,7 +21,7 @@ module.exports = function createUpdateEvent(event, github) {
         var file = new Buffer(JSON.stringify(event, null, 2)).toString('base64');
         var message = 'Updated event ' + previousEvent.id;
 
-        updateEventFile(github, config, commit.filename, eventsFile.sha, file, message)
+        githubHelper(github).updateEventFile(commit.filename, eventsFile.sha, file, message)
           .then(function() {
             resolve(event);
           })
@@ -41,7 +39,7 @@ module.exports = function createUpdateEvent(event, github) {
           var message = 'Created event ' + event.id;
 
           debug('GitHub:createFile()');
-          createEventFile(github, config, commit.filename, file, message)
+          githubHelper(github).createEventFile(commit.filename, file, message)
             .then(function() {
               return resolve(event);
             })
