+++ lib/talk/add-to-event-from-payload.js
@@ -1,9 +1,8 @@
 var config = require('../../common/config');
 var debug = require('debug')('gitevents-events');
-var getEventFile = require('../github/get-event-file');
+var githubHelper = require('../github-helper');
 var moment = require('moment');
 var S = require('string');
-var updateEventFile = require('../github/update-event-file');
 
 module.exports = function addTalkToEventFromPayload(payload, speaker, github) {
   return new Promise(function(resolve, reject) {
@@ -15,7 +14,7 @@ module.exports = function addTalkToEventFromPayload(payload, speaker, github) {
       filename: config.paths.events + eventId + '.json'
     };
 
-    getEventFile(github, config, commit.filename)
+    githubHelper(github).getEventFile(commit.filename)
       .then(function(eventsFile){
         var talks = [];
         var contents = new Buffer(eventsFile.content, eventsFile.encoding).toString('utf8');
@@ -50,7 +49,7 @@ module.exports = function addTalkToEventFromPayload(payload, speaker, github) {
         var file = new Buffer(JSON.stringify(event, null, 2)).toString('base64');
         var message = 'Updated event ' + event.id;
 
-        updateEventFile(github, config, commit.filename, eventsFile.sha, file, message)
+        githubHelper(github).updateEventFile(commit.filename, eventsFile.sha, file, message)
           .then(function() {
             return resolve(event);
           })
